refactor: combine duplicated function
