+++ lib/v2/pku/ss/notice.js
@@ -1,48 +1,10 @@
-const got = require('@/utils/got');
-const cheerio = require('cheerio');
-const { parseDate } = require('@/utils/parse-date');
+const { baseUrl, getSingleRecord, getArticle } = require('./common');
 
-const link = 'https://www.ss.pku.edu.cn';
-const host = 'https://www.ss.pku.edu.cn/newscenter/notice/';
-
-const getSingleRecord = async () => {
-    const res = await got(host);
-
-    const $ = cheerio.load(res.data);
-    const list = $('#info-list-ul').find('li');
-
-    return (
-        list &&
-        list
-            .map((index, item) => {
-                item = $(item);
-                const date = item.find('.time').text();
-                return {
-                    title: item.find('a').attr('title'),
-                    pubDate: parseDate(date),
-                    link: link + item.find('a').attr('href'),
-                };
-            })
-            .get()
-    );
-};
+const host = `${baseUrl}/newscenter/notice/`;
 
 module.exports = async (ctx) => {
-    const items = await getSingleRecord();
-    const out = await Promise.all(
-        items.map((item) =>
-            ctx.cache.tryGet(item.link, async () => {
-                const response = await got(item.link);
-                const $ = cheerio.load(response.data);
-                return {
-                    title: item.title,
-                    link: item.link,
-                    description: $('.article-content').html(),
-                    pubDate: item.pubDate,
-                };
-            })
-        )
-    );
+    const items = await getSingleRecord(host);
+    const out = await Promise.all(items.map((item) => getArticle(item, ctx.cache.tryGet)));
 
     ctx.state.data = {
         title: '北大软微-通知公告',
